Extract Substream struct to separate module
